add watcher option and client
